+++ service/controller/internal/cloudconfig/tccp_test.go
@@ -11,13 +11,13 @@ import (
 
 	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/pkg/apis/infrastructure/v1alpha2"
 	"github.com/giantswarm/certs"
-	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v_5_2_0"
+	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v6/v_6_0_0"
+	"github.com/giantswarm/micrologger/microloggertest"
 	"github.com/giantswarm/randomkeys"
 	"github.com/google/go-cmp/cmp"
 
 	"github.com/giantswarm/aws-operator/service/controller/internal/encrypter"
 	"github.com/giantswarm/aws-operator/service/controller/internal/unittest"
-	"github.com/giantswarm/micrologger/microloggertest"
 )
 
 var update = flag.Bool("update", false, "update .golden CF template file")
@@ -38,6 +38,7 @@ func Test_Controller_CloudConfig_TCCP_Template_Render(t *testing.T) {
 		ctx    context.Context
 		cr     infrastructurev1alpha2.AWSCluster
 		certs  certs.Cluster
+		images k8scloudconfig.Images
 		keys   randomkeys.Cluster
 		labels string
 	}{
@@ -46,6 +47,7 @@ func Test_Controller_CloudConfig_TCCP_Template_Render(t *testing.T) {
 			ctx:    unittest.DefaultContext(),
 			cr:     unittest.DefaultCluster(),
 			certs:  unittest.DefaultCerts(),
+			images: unittest.DefaultImages(),
 			keys:   unittest.DefaultKeys(),
 			labels: "k1=v1,k2=v2",
 		},
@@ -88,7 +90,7 @@ func Test_Controller_CloudConfig_TCCP_Template_Render(t *testing.T) {
 				}
 			}
 
-			templateBody, err := tccp.Render(tc.ctx, tc.cr, tc.certs, tc.keys, tc.labels)
+			templateBody, err := tccp.Render(tc.ctx, tc.cr, tc.certs, tc.keys, tc.images, tc.labels)
 			if err != nil {
 				t.Fatal(err)
 			}