+++ service/controller/internal/cloudconfig/template/aws-cni.go
@@ -1,4 +1,4 @@
-package cloudconfig
+package template
 
 const AwsCNIManifest = `
 # Vendored from https://raw.githubusercontent.com/aws/amazon-vpc-cni-k8s/master/config/v1.6/aws-k8s-cni.yaml