+++ service/controller/resource/drainfinisher/error.go
@@ -6,18 +6,6 @@ import (
 	"github.com/giantswarm/microerror"
 )
 
-// executionFailedError is an error type for situations where Resource execution
-// cannot continue and must always fall back to operatorkit.
-//
-// This error should never be matched against and therefore there is no matcher
-// implement. For further information see:
-//
-//     https://github.com/giantswarm/fmt/blob/master/go/errors.md#matching-errors
-//
-var executionFailedError = &microerror.Error{
-	Kind: "executionFailedError",
-}
-
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }
@@ -35,25 +23,25 @@ func IsMissingAnnotationError(err error) bool {
 	return microerror.Cause(err) == missingAnnotationError
 }
 
-var noActiveLifecycleActionError = &microerror.Error{
-	Kind: "noActiveLifecycleActionError",
+var noActiveLifeCycleActionError = &microerror.Error{
+	Kind: "noActiveLifeCycleActionError",
 }
 
-// IsNoActiveLifecycleAction asserts noActiveLifecycleActionError. It also
+// IsNoActiveLifeCycleAction asserts noActiveLifeCycleActionError. It also
 // checks for some string matching in the error message to figure if the AWS API
 // gives the error we expect.
-func IsNoActiveLifecycleAction(err error) bool {
+func IsNoActiveLifeCycleAction(err error) bool {
 	c := microerror.Cause(err)
 
 	if c == nil {
 		return false
 	}
 
-	if strings.Contains(c.Error(), "No active Lifecycle Action found") {
+	if strings.Contains(c.Error(), "no active life cycle action found") {
 		return true
 	}
 
-	if c == noActiveLifecycleActionError {
+	if c == noActiveLifeCycleActionError {
 		return true
 	}