+++ service/controller/resource/tccpazs/create_test.go
@@ -10,13 +10,13 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2"
 	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/pkg/apis/infrastructure/v1alpha2"
 	"github.com/giantswarm/apiextensions/pkg/clientset/versioned/fake"
-	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
-	"github.com/giantswarm/aws-operator/service/controller/internal/unittest"
-	"github.com/giantswarm/aws-operator/service/controller/key"
-	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger/microloggertest"
 	"github.com/giantswarm/to"
 	"github.com/google/go-cmp/cmp"
+
+	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
+	"github.com/giantswarm/aws-operator/service/controller/internal/unittest"
+	"github.com/giantswarm/aws-operator/service/controller/key"
 )
 
 func Test_EnsureCreated_AZ_Spec(t *testing.T) {
@@ -1036,10 +1036,6 @@ func Test_mapSubnets(t *testing.T) {
 			}
 		})
 	}
-}
-
-func isExecutionFailed(err error) bool {
-	return microerror.Cause(err) == executionFailedError
 }
 
 func mustParseCIDR(v string) net.IPNet {